src/templates/template-reinforcement-learning/requirements.txt
#::= from_template_common ::#
torchrl
moviepy